reworked tests, modified state_line
