--- tests/unit/analysis/___init__.py
@@ -1,5 +1,0 @@
-import importlib
-
-def test_import___init__():
-    # Smoke test: module should import successfully
-    importlib.import_module("part2pop.analysis.__init__")
