+++ tests/unit/population/factory/helpers/test_hiscale.py
@@ -1,0 +1,5 @@
+import importlib
+
+
+def test_import_population_factory_helper_hiscale():
+    importlib.import_module("part2pop.population.factory.helpers.hiscale")
