+++ tests/unit/viz/factory/test_series_line.py
