+++ tests/unit/viz/test_builder.py
@@ -1,7 +1,6 @@
 import pytest
 
 from part2pop.viz.builder import PlotBuilder, build_plotter
-from part2pop.viz.factory.series_line import SeriesLinePlotter
 
 
 def test_plot_builder_requires_type():
@@ -18,5 +17,5 @@ def test_plot_builder_unknown_type():
 def test_plot_builder_builds_registered_plotter():
     result = build_plotter("series_line", {"varname": "Nccn"})
 
-    assert isinstance(result, SeriesLinePlotter)
+    assert result.type == "series_line"
     assert result.config == {"varname": "Nccn"}
